Moved GOInvalidFile and GOStandardFile to the files subdirectory
