+++ src/core/files/GOInvalidFile.cpp
