+++ src/grandorgue/GOOrganController.cpp
@@ -29,6 +29,7 @@
 #include "control/GOElementCreator.h"
 #include "control/GOPushbuttonControl.h"
 #include "dialogs/GOProgressDialog.h"
+#include "files/GOOpenedFile.h"
 #include "files/GOStdFileName.h"
 #include "gui/GOGUIBankedGeneralsPanel.h"
 #include "gui/GOGUICouplerPanel.h"
@@ -65,7 +66,6 @@
 #include "GODivisionalCoupler.h"
 #include "GODocument.h"
 #include "GOEvent.h"
-#include "GOFile.h"
 #include "GOHash.h"
 #include "GOMetronome.h"
 #include "GOOrgan.h"
